requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.1.0
bleach==6.0.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.2.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-aiplatform==1.25.0
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-generativeai==0.1.0rc2
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.54.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
langchain==0.0.173
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.2
openai==0.27.2
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.6
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2022.10.31
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
rsa==4.9
semver==2.13.0
Send2Trash==1.8.0
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
stack-data==0.6.2
streamlit==1.22.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
tiktoken==0.3.2
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
yarl==1.8.2
zipp==3.15.0
